Improved frontier identification.
Before, since the point I'm searching around is in the center of a square
and I was in a hurry, I opted to create four for loops, each one searching
one quadrant of the square. Now, just one for loop is used to comb through
the entirety of the square.